+++ include/terrain.h
@@ -100,7 +100,7 @@ private:
 	double moveFraction;
 	bool disabled;
 	Object *lastEntered;
-	long lastRender;
+	Uint32 lastRender;
 public:
 	int dir;
 	bool requestEntry(Object* other, int dir);
